# project-roadmap/signalarc-custom-gpt-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SignalArc Custom GPT Action API",
    "version": "0.2.0",
    "description": "OpenAPI schema for the SignalArc Custom GPT action surface. Covers Circle Agent Wallet onboarding, sessions, wallet/balance visibility, market intent lifecycle, and read-only market discovery on Arc Testnet. Custom GPT MUST always call a matching Action when the user asks to onboard, verify, register, check session/wallet/balance, request faucet, list markets, or create/confirm/execute an intent. The GPT MUST generate a unique agent_id for each new user and MUST NOT use generic placeholder agent_id values such as signalarc-gpt-agent, agent_desi_001, default, test, demo, user, or chatgpt. The recommended agent_id shape is agent_<slug>_chatgpt_<suffix>, for example agent_sanatarau21_chatgpt_001. For create_market intents, close_timestamp MUST be sent as a UTC RFC3339 string such as 2026-05-31T23:59:00Z or as a unix-seconds integer string such as 1780000000; natural-language dates such as 'end of May 2026' MUST be converted before the Action is called."
  },
  "servers": [
    {
      "url": "https://api.signalarc.fun"
    }
  ],
  "paths": {
...
    "/markets": {
      "get": {
        "operationId": "listMarkets",
        "summary": "List up to 50 SignalArc markets for read-only discovery",
        "description": "Read-only market discovery. Returns up to 50 SignalArc markets on Arc Testnet with status (DRAFT, OPEN, CLOSED, RESOLVED, SETTLED), category, collateral_asset, chain, and closes_at. Call this Action when the user asks for open markets, trending markets, or any market list.",
        "responses": {
          "200": {
            "description": "Markets list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "markets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Market"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Markets list failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/{market_id}": {
      "get": {
        "operationId": "getMarket",
        "summary": "Get a SignalArc market by id",
        "description": "Returns a single SignalArc market record by id. Read-only.",
        "parameters": [
          {
            "name": "market_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Market UUID returned by listMarkets."
          }
        ],
        "responses": {
          "200": {
            "description": "Market record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "market": {
                      "$ref": "#/components/schemas/Market"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Market not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/markets": {
      "get": {
        "operationId": "listAgentMarkets",
        "summary": "List agent-readable SignalArc markets",
        "description": "Compact, agent-friendly market list for SignalArc Arc Testnet markets. Same dataset as listMarkets but with a flat shape suited to natural-language summaries. Read-only. The Custom GPT SHOULD call this Action when the user asks to see open markets, trending markets, or any market list.",
        "responses": {
          "200": {
            "description": "Agent markets list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "markets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AgentMarket"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Agent markets list failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Market": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "creator_user_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "outcome_yes_label": {
            "type": "string"
          },
          "outcome_no_label": {
            "type": "string"
          },
          "collateral_asset": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "resolution_source": {
            "type": "string"
          },
          "opens_at": {
            "type": "string"
          },
          "closes_at": {
            "type": "string"
          },
          "resolved_at": {
            "type": "string"
          },
          "settled_at": {
            "type": "string"
          },
          "winning_outcome": {
            "type": "string"
          },
          "market_contract_address": {
            "type": "string"
          },
          "market_deployment_tx_hash": {
            "type": "string"
          },
          "market_factory_address": {
            "type": "string"
          },
          "resolver_address": {
            "type": "string"
          },
          "onchain_deployment_status": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "AgentMarket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "collateral_asset": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "closes_at": {
            "type": "string"
          },
          "market_contract_address": {
            "type": "string"
          },
          "resolution_source": {
            "type": "string"
          }
        }
      },
...
    }
  }
}